move string conversion
