+++ lib/ethpy/ethpy/hyperdrive/__init__.py
@@ -1,5 +1,5 @@
 """Interfaces for bots and hyperdrive smart contracts."""
-from .addresses import HyperdriveAddresses, camel_to_snake, fetch_hyperdrive_address_from_uri
+from .addresses import HyperdriveAddresses, fetch_hyperdrive_address_from_uri
 from .assets import BASE_TOKEN_SYMBOL, AssetIdPrefix, decode_asset_id, encode_asset_id
 from .deploy import DeployedHyperdrivePool, deploy_hyperdrive_from_factory
 from .errors import HyperdriveErrors, lookup_hyperdrive_error_selector
