+++ lib/ethpy/ethpy/hyperdrive/state/conversions.py
@@ -1,10 +1,10 @@
 """Functions for converting Hyperdrive state values."""
 from __future__ import annotations
 
+import re
 from dataclasses import asdict
 from typing import Any
 
-from ethpy.hyperdrive.addresses import camel_to_snake, snake_to_camel
 from fixedpointmath import FixedPoint
 from hypertypes.IHyperdriveTypes import Checkpoint as HtCheckpoint
 from hypertypes.IHyperdriveTypes import Fees as HtFees
@@ -15,6 +15,19 @@ from .checkpoint import Checkpoint
 from .fees import Fees
 from .pool_config import PoolConfig
 from .pool_info import PoolInfo
+
+
+def camel_to_snake(snake_string: str) -> str:
+    """Convert camel case string to snake case string."""
+    return re.sub(r"(?<!^)(?=[A-Z])", "_", snake_string).lower()
+
+
+def snake_to_camel(snake_string: str) -> str:
+    """Convert snake case string to camel case string."""
+    # First capitalize the letters following the underscores and remove underscores
+    camel_string = re.sub(r"_([a-z])", lambda x: x.group(1).upper(), snake_string)
+    # Ensure the first character is lowercase to achieve lowerCamelCase
+    return camel_string[0].lower() + camel_string[1:] if camel_string else camel_string
 
 
 def dataclass_to_dict(
